requirements.txt
lark-parser
PyQt5